lib: Updated API
